+++ plugins/css-prefers-color-scheme/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to Prefers Color Scheme
 
+### Unreleased
+
+- Improve `types` declaration in `package.json`
+
 ### 8.0.0 (January 24, 2023)
 
 - Updated: Support for Node v14+ (major).