+++ plugins/postcss-attribute-case-insensitive/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Attribute Case Insensitive
 
+### Unreleased
+
+- Improve `types` declaration in `package.json`
+
 ### 6.0.0 (January 24, 2023)
 
 - Updated: Support for Node v14+ (major).