+++ plugins/postcss-logical-resize/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Logical Resize
 
+### Unreleased
+
+- Improve `types` declaration in `package.json`
+
 ### 1.0.0 (January 21, 20223)
 
 - Initial version