+++ plugins/postcss-nesting/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Nesting
 
+### Unreleased
+
+- Improve `types` declaration in `package.json`
+
 ### 11.0.0 (January 24, 2023)
 
 - Updated: Support for Node v14+ (major).