+++ plugins/postcss-pseudo-class-any-link/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Pseudo Class Any Link
 
+### Unreleased
+
+- Improve `types` declaration in `package.json`
+
 ### 8.0.0 (January 24, 2023)
 
 - Updated: Support for Node v14+ (major).